requirements.txt
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828496740/work
asgiref==3.2.10
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
certifi==2020.6.20
cffi==1.14.0
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1602083700509/work
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
entrypoints==0.3
image==1.5.32
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
imutils==0.5.3
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils==0.2.0
jedi @ file:///tmp/build/80754af9/jedi_1598371618777/work
Jinja2==2.11.2
jsonschema @ file:///tmp/build/80754af9/jsonschema_1601490610665/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1601059699549/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914804165/work
nbformat==5.0.7
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook @ file:///tmp/build/80754af9/notebook_1601501581489/work
numpy==1.19.2
olefile==0.46
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pdf2image==1.14.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1602788709045/work
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1598885458782/work
ptyprocess==0.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141725711/work
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658315997/work
PyYAML==5.3.1
pyzmq==19.0.2
scikit-image==0.16.2
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
tornado==6.0.4
traitlets==4.3.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
zipp @ file:///tmp/build/80754af9/zipp_1602002379921/work
